commit : * move ssl implementation to `blynklib_cp.py` * move cert to certificate folder
